requirements.txt
pandas==2.0.2
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
Flask==2.0.1
Flask-PyMongo==2.3.0
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
openai==0.27.7
langchain==0.0.183
streamlit==1.22.0
streamlit-chat==0.0.2.2
gradio==3.32.0
six==1.16.0
urllib3==1.26.5